backend/requirements.txt
bcrypt==5.0.0
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
Flask==3.1.3
flask-cors==6.0.2
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
marshmallow==4.3.0
mysql-connector-python==9.6.0
PyJWT==2.12.1
python-dotenv==1.2.2
requests==2.33.1
stripe==15.0.1
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.8
